Adjust CombineZernikesSigmaClipTask methods.

+++ python/lsst/ts/wep/task/CombineZernikesMeanTask.py
@@ -32,4 +32,4 @@ class CombineZernikesMeanTask(CombineZernikesBaseTask):
 
     def combineZernikes(self, zernikeArray):
 
-        return np.mean(zernikeArray, axis=0), np.zeros(len(zernikeArray))
+        return np.mean(zernikeArray, axis=0), np.zeros(len(zernikeArray), dtype=int)
